Move time utils to `:shared:base`
