--- shared/base/src/main/java/com/ivy/base/TimeProvider.kt
@@ -1,9 +1,0 @@
-package com.ivy.base
-
-import java.time.ZoneId
-import javax.inject.Inject
-
-@Suppress("UnnecessaryPassThroughClass")
-class TimeProvider @Inject constructor() {
-    fun getZoneId(): ZoneId = ZoneId.systemDefault()
-}
