+++ shared/base/src/main/java/com/ivy/base/di/BaseHiltBindings.kt
@@ -2,6 +2,10 @@ package com.ivy.base.di
 
 import com.ivy.base.threading.DispatchersProvider
 import com.ivy.base.threading.IvyDispatchersProvider
+import com.ivy.base.time.TimeConverter
+import com.ivy.base.time.TimeProvider
+import com.ivy.base.time.impl.DeviceTimeProvider
+import com.ivy.base.time.impl.StandardTimeConvert
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -9,7 +13,13 @@ import dagger.hilt.components.SingletonComponent
 
 @Module
 @InstallIn(SingletonComponent::class)
-abstract class BaseHiltBindings {
+interface BaseHiltBindings {
     @Binds
-    abstract fun dispatchersProvider(impl: IvyDispatchersProvider): DispatchersProvider
+    fun dispatchersProvider(impl: IvyDispatchersProvider): DispatchersProvider
+
+    @Binds
+    fun bindTimezoneProvider(impl: DeviceTimeProvider): TimeProvider
+
+    @Binds
+    fun bindTimeConverter(impl: StandardTimeConvert): TimeConverter
 }
