+++ shared/base/src/main/java/com/ivy/base/time/TimeProvider.kt
@@ -1,0 +1,7 @@
+package com.ivy.base.time
+
+import java.time.ZoneId
+
+interface TimeProvider {
+    fun getZoneId(): ZoneId
+}
