+++ shared/base/src/main/java/com/ivy/base/time/impl/DeviceTimeProvider.kt
@@ -1,0 +1,11 @@
+package com.ivy.base.time.impl
+
+import com.ivy.base.time.TimeProvider
+import java.time.ZoneId
+import javax.inject.Inject
+
+class DeviceTimeProvider @Inject constructor() : TimeProvider {
+    override fun getZoneId(): ZoneId {
+        return ZoneId.systemDefault()
+    }
+}
