+++ shared/data/core/src/test/java/com/ivy/data/repository/mapper/TransactionMapperTest.kt
@@ -3,8 +3,8 @@ package com.ivy.data.repository.mapper
 import arrow.core.Some
 import com.google.testing.junit.testparameterinjector.TestParameter
 import com.google.testing.junit.testparameterinjector.TestParameterInjector
-import com.ivy.base.TimeProvider
 import com.ivy.base.model.TransactionType
+import com.ivy.base.time.TimeProvider
 import com.ivy.data.db.entity.TransactionEntity
 import com.ivy.data.model.AccountId
 import com.ivy.data.model.CategoryId
