+++ shared/domain/src/main/java/com/ivy/domain/di/IvyCoreBindingsModule.kt
@@ -2,10 +2,6 @@ package com.ivy.domain.di
 
 import com.ivy.domain.features.Features
 import com.ivy.domain.features.IvyFeatures
-import com.ivy.domain.time.TimeConverter
-import com.ivy.domain.time.TimeZoneProvider
-import com.ivy.domain.time.impl.DeviceTimeZoneProvider
-import com.ivy.domain.time.impl.StandardTimeConvert
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -16,10 +12,4 @@ import dagger.hilt.components.SingletonComponent
 interface IvyCoreBindingsModule {
     @Binds
     fun bindFeatures(features: IvyFeatures): Features
-
-    @Binds
-    fun bindTimezoneProvider(impl: DeviceTimeZoneProvider): TimeZoneProvider
-
-    @Binds
-    fun bindTimeConverter(impl: StandardTimeConvert): TimeConverter
 }
