--- shared/domain/src/main/java/com/ivy/domain/time/TimeZoneProvider.kt
@@ -1,7 +1,0 @@
-package com.ivy.domain.time
-
-import java.time.ZoneId
-
-interface TimeZoneProvider {
-    fun getZoneId(): ZoneId
-}
