--- shared/domain/src/main/java/com/ivy/domain/time/impl/DeviceTimeZoneProvider.kt
@@ -1,11 +1,0 @@
-package com.ivy.domain.time.impl
-
-import com.ivy.domain.time.TimeZoneProvider
-import java.time.ZoneId
-import javax.inject.Inject
-
-class DeviceTimeZoneProvider @Inject constructor() : TimeZoneProvider {
-    override fun getZoneId(): ZoneId {
-        return ZoneId.systemDefault()
-    }
-}
